+++ packages/antd/src/OneOfAnt.tsx
@@ -1,4 +1,4 @@
-import { idToDomId, OneOf } from '@moxb/moxb';
+import { BindOneOfChoice, idToDomId, OneOf } from '@moxb/moxb';
 import { Button, Dropdown, Menu, Radio, Select } from 'antd';
 import { DropDownProps } from 'antd/lib/dropdown';
 import { FormItemProps } from 'antd/lib/form/FormItem';
@@ -224,6 +224,57 @@ export class OneOfSelectAnt extends React.Component<BindAntProps<OneOf> & Select
                 {...props}
             >
                 {operation.choices.map((opt) => (
+                    <Select.Option
+                        data-testid={idToDomId(operation.id + '.' + opt.value)}
+                        key={opt.value}
+                        value={opt.value}
+                    >
+                        {opt.widget ? opt.widget : opt.label}
+                    </Select.Option>
+                ))}
+            </Select>
+        );
+    }
+}
+
+interface OneOfSearchableSelectAntState<T = string> {
+    value?: string;
+    data: BindOneOfChoice<T>[];
+}
+
+@observer
+export class OneOfSearchableSelectAnt extends React.Component<
+    BindAntProps<OneOf> & SelectProps<any>,
+    OneOfSearchableSelectAntState<any>
+> {
+    render() {
+        const { operation, invisible, readOnly, mode, ...props } = parseProps(this.props, this.props.operation);
+
+        if (invisible) {
+            return null;
+        }
+        if (readOnly) {
+            return <span data-testid={operation.id}>{operation.choice}</span>;
+        }
+        return (
+            <Select
+                showSearch
+                defaultActiveFirstOption={false}
+                showArrow={false}
+                data-testid={operation.id}
+                onChange={(selectionValue: any) => {
+                    operation.setValue(selectionValue);
+                    operation.searchData('');
+                }}
+                onSearch={operation.searchData}
+                notFoundContent={null}
+                filterOption={false}
+                // mode={mode}
+                value={operation.value || undefined}
+                placeholder={operation.placeholder}
+                {...props}
+            >
+                {operation.filteredChoices.map((opt: BindOneOfChoice<any>) => (
                     <Select.Option
                         data-testid={idToDomId(operation.id + '.' + opt.value)}
                         key={opt.value}