+++ packages/moxb/src/one-of/OneOf.ts
@@ -22,4 +22,6 @@ export interface BindOneOfChoice<T = string> {
 export interface OneOf<T = string> extends Value<T> {
     readonly choices: BindOneOfChoice<T>[];
     readonly choice: string | undefined;
+    searchData(value: string): void;
+    readonly filteredChoices: BindOneOfChoice<T>[];
 }